Online Server list Alpha :green_book:
Now everytime we launch the server we emit that it is online and it will now show up inside the server list! Awesome! :heart:

Installed UUID because we want to have ids for servers that can be useful later.

When you exit the window please make sure that you press "ctrl + c". Then we can emit that the server is shutdown to the main server.

SOON: I'm going to make a system if server owner refuses to use ctrl + c in order to shutdown the server. It will work something like this:

- MainServer: Hello? You still alive?
- CommunityServer: Yes, I'm here.
(15 minutes of silence)
- MainServer: Hello? You still alive?
- CommunityServer: Yes, I'm here.
(15 minutes of silence)
- MainServer: Hello? You still alive?
(No answer)

:heart: